pyproject.toml
[tool.poetry]
name = "hotel-email-parser"
version = "0.1.0"
description = "Production-grade hotel booking email parsing system with hybrid NLP architecture"
authors = ["Applied NLP Team"]
readme = "README.md"
packages = [{include = "pipeline"}, {include = "utils"}]

[tool.poetry.dependencies]
python = "^3.10"
spacy = "^3.7"
transformers = "^4.36"
sentence-transformers = "^2.2"
torch = "^2.1"
dateparser = "^1.2"
scikit-learn = "^1.3"
pandas = "^2.1"
PyYAML = "^6.0"
numpy = "^1.26"
matplotlib = "^3.8"
seaborn = "^0.13"
click = "^8.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
black = "^23.12"
mypy = "^1.7"
flake8 = "^6.1"
ipython = "^8.18"

[tool.poetry.scripts]
parse-email = "parse_email:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
